rename dash_windows, use lookat for notify panel floating
