+++ scripts/system/dashboard/windows.js
